Single responsibility (connect)
